+++ .eslintrc.cjs
@@ -32,5 +32,12 @@ module.exports = {
         // 'no-undef': 0,
         'no-new-object': 1, // 使用字面量创建对象 非new Object()
         'object-shorthand': 1, // 对象方法属性值缩写
+        'vue/attribute-hyphenation': [
+            'warn',
+            'never',
+            {
+                ignore: ['custom-prop'],
+            },
+        ],
     },
 };